Use storage.SeriesRef in active series tracker (#5664)
* Use storage.SeriesRef in active series tracker

We're not tracking generic uint64 but a storage.SeriesRef.
This isn't changing any logic.


* Fix one more test


---------

+++ pkg/ingester/ingester.go
@@ -1098,7 +1098,7 @@ func (i *Ingester) pushSamplesToAppender(userID string, timeseries []mimirpb.Pre
 		}
 
 		if activeSeries != nil && stats.succeededSamplesCount > oldSucceededSamplesCount {
-			activeSeries.UpdateSeries(nonCopiedLabels, uint64(ref), startAppend, numNativeHistogramBuckets)
+			activeSeries.UpdateSeries(nonCopiedLabels, ref, startAppend, numNativeHistogramBuckets)
 		}
 
 		if len(ts.Exemplars) > 0 && i.limits.MaxGlobalExemplarsPerUser(userID) > 0 {
